Move LeaderboardScoreSortOrder under RiderStandingTransformation
